introduce new hook useCommandMenuHistory
